Un-nest private implentations of IFileInfo and IDirectoryContents
Fixes #156
