+++ src/Microsoft.AspNet.FileProviders.Composite/CompositeDirectoryContents.cs
@@ -5,10 +5,11 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 
-namespace Microsoft.AspNet.FileProviders
+namespace Microsoft.AspNet.FileProviders.Composite
 {
     /// <summary>
-    /// Represents the result of a call composition of <see cref="IFileProvider.GetDirectoryContents(string)"/> for a list of <see cref="IFileProvider"/> and a path.
+    /// Represents the result of a call composition of <see cref="IFileProvider.GetDirectoryContents(string)"/>
+    /// for a list of <see cref="IFileProvider"/> and a path.
     /// </summary>
     public class CompositeDirectoryContents : IDirectoryContents
     {
@@ -19,13 +20,14 @@ namespace Microsoft.AspNet.FileProviders
         private List<IDirectoryContents> _directories;
 
         /// <summary>
-        /// Creates a new instance of <see cref="CompositeDirectoryContents"/> to represents the result of a call composition of <see cref="IFileProvider.GetDirectoryContents(string)"/>.
+        /// Creates a new instance of <see cref="CompositeDirectoryContents"/> to represents the result of a call composition of
+        /// <see cref="IFileProvider.GetDirectoryContents(string)"/>.
         /// </summary>
         /// <param name="fileProviders">The list of <see cref="IFileProvider"/> for which the results have to be composed.</param>
         /// <param name="subpath">The path.</param>
         public CompositeDirectoryContents(IList<IFileProvider> fileProviders, string subpath)
         {
-            if(fileProviders == null)
+            if (fileProviders == null)
             {
                 throw new ArgumentNullException(nameof(fileProviders));
             }
