+++ src/Microsoft.AspNet.FileProviders.Embedded/EmbeddedFileProvider.cs
@@ -7,6 +7,7 @@ using System.IO;
 using System.Linq;
 using System.Reflection;
 using System.Text;
+using Microsoft.AspNet.FileProviders.Embedded;
 using Microsoft.Extensions.Primitives;
 
 namespace Microsoft.AspNet.FileProviders
@@ -96,6 +97,7 @@ namespace Microsoft.AspNet.FileProviders
             {
                 return new NotFoundFileInfo(name);
             }
+
             return new EmbeddedResourceFileInfo(_assembly, resourcePath, name, _lastModified);
         }
 
@@ -153,62 +155,6 @@ namespace Microsoft.AspNet.FileProviders
         private static bool HasInvalidPathChars(string path)
         {
             return path.IndexOfAny(_invalidFileNameChars) != -1;
-        }
-
-        private class EmbeddedResourceFileInfo : IFileInfo
-        {
-            private readonly Assembly _assembly;
-            private readonly string _resourcePath;
-
-            private long? _length;
-
-            public EmbeddedResourceFileInfo(
-                Assembly assembly,
-                string resourcePath,
-                string name,
-                DateTimeOffset lastModified)
-            {
-                _assembly = assembly;
-                _resourcePath = resourcePath;
-                Name = name;
-                LastModified = lastModified;
-            }
-
-            public bool Exists => true;
-
-            public long Length
-            {
-                get
-                {
-                    if (!_length.HasValue)
-                    {
-                        using (var stream = _assembly.GetManifestResourceStream(_resourcePath))
-                        {
-                            _length = stream.Length;
-                        }
-                    }
-                    return _length.Value;
-                }
-            }
-
-            // Not directly accessible.
-            public string PhysicalPath => null;
-
-            public string Name { get; }
-
-            public DateTimeOffset LastModified { get; }
-
-            public bool IsDirectory => false;
-
-            public Stream CreateReadStream()
-            {
-                var stream = _assembly.GetManifestResourceStream(_resourcePath);
-                if (!_length.HasValue)
-                {
-                    _length = stream.Length;
-                }
-                return stream;
-            }
         }
     }
 }
