+++ src/Microsoft.AspNet.FileProviders.Sources/Constants.cs
