+++ src/Microsoft.AspNet.FileProviders.Sources/EmptyDisposable.cs
