+++ src/Microsoft.AspNet.FileProviders.Sources/EnumerableDirectoryContents.cs
