+++ src/Microsoft.AspNet.FileProviders.Sources/NoopChangeToken.cs
