+++ src/Microsoft.AspNet.FileProviders.Sources/NotFoundFileInfo.cs
