+++ test/Microsoft.AspNet.FileProviders.Composite.Tests/MockDisposable.cs
@@ -3,7 +3,7 @@
 
 using System;
 
-namespace Microsoft.AspNet.FileProviders.Composite.Tests.TestUtility
+namespace Microsoft.AspNet.FileProviders.Composite
 {
     public class MockDisposable : IDisposable
     {
