+++ test/Microsoft.AspNet.FileProviders.Physical.Tests/project.json
@@ -3,21 +3,17 @@
     "warningsAsErrors": true,
     "keyFile": "../../tools/Key.snk"
   },
-    "dependencies": {
-        "Microsoft.AspNet.FileProviders.Physical": "1.0.0-*",
-        "Microsoft.AspNet.Testing": "1.0.0-*",
-        "Microsoft.Extensions.FileSystemGlobbing.Tests": "1.0.0-*",
-        "xunit.runner.aspnet": "2.0.0-aspnet-*"
-    },
-    "commands": {
-        "test": "xunit.runner.aspnet"
-    },
-    "frameworks": {
-        "dnx451": { },
-        "dnxcore50": {
-            "dependencies": {
-                "System.Threading.Tasks.Parallel": "4.0.1-*"
-            }
-        }
-    }
+  "compile": "../Microsoft.Extensions.FileSystemGlobbing.Tests/TestUtility/DisposableFileSystem.cs",
+  "dependencies": {
+    "Microsoft.AspNet.FileProviders.Physical": "1.0.0-*",
+    "Microsoft.AspNet.Testing": "1.0.0-*",
+    "xunit.runner.aspnet": "2.0.0-aspnet-*"
+  },
+  "commands": {
+    "test": "xunit.runner.aspnet"
+  },
+  "frameworks": {
+    "dnx451": { },
+    "dnxcore50": { }
+  }
 }
